convert into function component MovieList
