+++ src/components/MovieList/MovieList.module.css
@@ -1,0 +1,7 @@
+.movie_list {
+  display: flex;
+  flex-wrap: wrap;
+  list-style: none;
+  gap: 20px;
+  justify-content: center;
+}
